Extract yo method
phovea/generator-phovea#425
